Relocate platform partials under top-level folders
